services/actual-budget/requirements.txt
actualpy>=0.9.0
httpx>=0.26.0
python-dotenv>=1.0.0
